test: Test router Code
