+++ js/main.js
@@ -8,6 +8,7 @@ import { authLogin, editUser } from "./requests.js";
 import { createItemEvent, renderAdminItems } from "./admin.js";
 import { getItem } from "./requests.js";
 import { render, sassFalse } from "sass";
+import { getMainPage } from "./render.js";
 
 // 관리자 이메일 -> 추후 .env넣어야함.
 const ADMIN_EMAIL = `hyochofriend@naver.com`;
@@ -34,6 +35,7 @@ export const loginPw = document.querySelector(".login-pw");
 export const loginBtn = document.querySelector(".login-btn");
 export const idboxEl = document.querySelector(".id-box");
 export const pwboxEl = document.querySelector(".pw-box");
+export const loginErrorBox = document.querySelector(".login-error-box");
 
 //search elements
 const searchInput = document.getElementById("search-main");
@@ -53,12 +55,6 @@ const pwChangeBtn = document.querySelector(".pw-change-btn");
 
 //상세페이지
 const detailPageEl = document.querySelector(".detail-container");
-
-//메인페이지
-const itemimgEl = document.querySelector(".itemlist-image");
-const itemTagsEls = document.querySelectorAll(".itemlist-tag");
-const itemTitleEls = document.querySelectorAll(".itemlist-title");
-const itemPriceEls = document.querySelectorAll(".itemlist-price");
 
 firstNav.addEventListener("mouseover", () => {
   backGround.style.visibility = "visible";
@@ -86,7 +82,9 @@ loginBtnEl.addEventListener("click", () => {
     });
   }
 });
-
+// 평소에는 display: none을 걸어놓는다.
+// 이름 변경을 클릭하고 완료 했으면 모달창이 뜨도록
+// 확인 버튼을 누르면 다시 display: none 되도록
 export const userModal = document.querySelector(".user-modal");
 const userModalBtn = document.querySelector(".user-modal-btn");
 export const userModalContent = document.querySelector(".user-modal-content");
@@ -125,39 +123,6 @@ pwChangeBtn.addEventListener("click", async (event) => {
   // 만료시간 체크는 계속
   getItemWithExpireTime("token");
 })();
-
-// 로그인 시 유효성 검사
-const idErrorMsg = document.querySelector(".id-error-msg");
-loginId.addEventListener("focusout", () => {
-  const exptext = /^[A-Za-z0-9_\.\-]+@[A-Za-z0-9\-]+\.[A-Za-z0-9\-]+/;
-  if (loginId.value && !exptext.test(loginId.value)) {
-    idErrorMsg.classList.add("show");
-    idboxEl.style.border = "1px solid #ed234b";
-  }
-});
-loginId.addEventListener("focusin", () => {
-  idErrorMsg.classList.remove("show");
-  idboxEl.style.border = "1px solid #999";
-});
-
-// 로그인 실패 시
-const loginErrorBox = document.querySelector(".login-error-box");
-export function showErrorBox() {
-  loginErrorBox.classList.add("show");
-  setTimeout(() => {
-    loginErrorBox.classList.remove("show");
-  }, 2000);
-}
-
-// 회원가입 유효성 검사
-// const singupEmailBox = document.querySelector('.signup-email-box')
-// emailInputEl.addEventListener("focusout", () => {
-//   const exptext = /^[A-Za-z0-9_\.\-]+@[A-Za-z0-9\-]+\.[A-Za-z0-9\-]+/;
-//   if (emailInputEl.value && !exptext.test(emailInputEl)) {
-//     // idErrorMsg.classList.add("show");
-//     singupEmailBox.style.border = "1px solid #ed234b";
-//   }
-// });
 
 // ============ 관리자페이지 ============
 let base64Thumbnail = "";
@@ -240,10 +205,7 @@ userInfoBtn.addEventListener("click", () => {
   const token = localStorage.getItem("token");
   if (token) {
     window.location = "#/user";
-  } else {
-    userModalContent.innerHTML = `로그인을 해주세요.`;
-    userModal.classList.add("show");
-  }
+  } else return;
 });
 
 // bank elements
@@ -327,23 +289,4 @@ bankSubmitBtn.addEventListener("click", () => {
   bankNumber = "";
 });
 
-async function getMainPage() {
-  const items = await getItem();
-  const christmasItem = items.filter((item) => item.tags[0] === "크리스마스");
-  const planterior = items.filter((item) => item.tags[0] === "플랜테리어");
-  const cookoo = items.filter((item) => item.tags[0] === "쿠쿠");
-  const drawer = items.filter((item) => item.tags[0] === "플랜테리어");
-  for (let i = 0; i < 4; i++) {
-    itemTagsEls[i].innerHTML = christmasItem[i].tags;
-    itemTitleEls[i].innerHTML = christmasItem[i].title;
-    itemPriceEls[i].innerHTML = christmasItem[i].price;
-  }
-  for (let i = 4; i < 8; i++) {
-    itemTagsEls[i].innerHTML = planterior[i].tags;
-    itemTitleEls[i].innerHTML = planterior[i].title;
-    itemPriceEls[i].innerHTML = planterior[i].price;
-  }
-}
 getMainPage();
-
-console.log(await getItem());
