+++ js/store.js
@@ -1,0 +1,5 @@
+//메인페이지
+export const itemimgEl = document.querySelector(".itemlist-image");
+export const itemTagsEls = document.querySelectorAll(".itemlist-tag");
+export const itemTitleEls = document.querySelectorAll(".itemlist-title");
+export const itemPriceEls = document.querySelectorAll(".itemlist-price");
